fix: combine post-build scripts
